+++ common/cpp/s3_rate_limiter.h
@@ -17,28 +17,40 @@
 
 #pragma once
 
-#include <bvar/bvar.h>
-
 #include <chrono>
+#include <functional>
 #include <memory>
 #include <shared_mutex>
 
-#include "common/status.h"
-#include "util/spinlock.h"
 namespace doris {
-
-enum class S3RateLimitType : size_t {
+enum class S3RateLimitType : int {
     GET = 0,
     PUT,
     UNKNOWN,
 };
+
 extern std::string to_string(S3RateLimitType type);
 extern S3RateLimitType string_to_s3_rate_limit_type(std::string_view value);
-extern Status reset_s3_rate_limiter(S3RateLimitType type, size_t max_speed, size_t max_burst,
-                                    size_t limit);
+
+class SpinLock {
+public:
+    SpinLock() = default;
+
+    void lock() {
+        while (_flag.test_and_set(std::memory_order_acq_rel)) {
+            // Spin until we acquire the lock
+        }
+    }
+
+    void unlock() { _flag.clear(std::memory_order_acq_rel); }
+
+private:
+    std::atomic_flag _flag = ATOMIC_FLAG_INIT;
+};
+
 class S3RateLimiter {
 public:
-    static const size_t default_burst_seconds = 1;
+    static constexpr size_t default_burst_seconds = 1;
 
     S3RateLimiter(size_t max_speed, size_t max_burst, size_t limit)
             : _max_speed(max_speed),
@@ -58,22 +70,25 @@ private:
     const size_t _max_burst {0}; // in tokens. which indicates the token bucket size
     const uint64_t _limit {0};   // 0 - not limited.
     SpinLock _mutex;
-    double _remain_tokens {
-            0}; // Amount of remain_tokens available in token bucket. Updated in `add` method.
+    // Amount of remain_tokens available in token bucket. Updated in `add` method.
+    double _remain_tokens {0};
     std::chrono::system_clock::time_point _prev_ms; // Previous `add` call time (in nanoseconds).
 };
 
 class S3RateLimiterHolder {
 public:
-    S3RateLimiterHolder(S3RateLimitType type, size_t max_speed, size_t max_burst, size_t limit);
+    S3RateLimiterHolder(S3RateLimitType type, size_t max_speed, size_t max_burst, size_t limit,
+                        std::function<void(int64_t)> metric_func);
 
     int64_t add(size_t amount);
 
-    Status reset(size_t max_speed, size_t max_burst, size_t limit);
+    int reset(size_t max_speed, size_t max_burst, size_t limit);
 
 private:
     std::shared_mutex rate_limiter_rw_lock;
     std::unique_ptr<S3RateLimiter> rate_limiter;
-    bvar::Adder<uint64_t> rate_limit_bvar;
+    // Record the correspoding sleeping time(unit is ms)
+    std::function<void(int64_t)> metric_func;
 };
+
 } // namespace doris
